Module type Sig.BITVECTOR_LATTICE_NO_CONVERSION

Bitvector Lattice: a LATTICE representing fixed sized bitvectors. The size isn't stored in the lattice, instead lattice operations take it as an extra argument.

include Datatype_sig.S

type t

val equal : t -> t -> bool

Any notion of equality is allowed, as long as it is an equivalence relation, and that if a == b, then equal a b.

val compare : t -> t -> int

compare is a total order, and should be compatible with equal.

val hash : t -> int

hash requires that equal values have the same hash.

val bottom : size:Units.In_bits.t -> t

val is_bottom : size:Units.In_bits.t -> t -> bool

val top : size:Units.In_bits.t -> t

val inter : size:Units.In_bits.t -> t -> t -> t

val join : size:Units.In_bits.t -> t -> t -> t

val pretty : size:Units.In_bits.t -> Format.formatter -> t -> unit

val widen : size:Units.In_bits.t -> previous:t -> t -> t

val includes : size:Units.In_bits.t -> t -> t -> bool

val includes_or_widen : size:Units.In_bits.t -> previous:t -> t -> bool * t

val singleton : size:Units.In_bits.t -> Z.t -> t