Module type Sig.BITVECTOR_QUERIES

type t

val is_singleton : size:Units.In_bits.t -> t -> Z.t option

val is_empty : size:Units.In_bits.t -> t -> bool

True if the binary cannot be concretized into any value.

val fold_crop_signed : size:Units.In_bits.t -> t -> inf:Z.t -> sup:Z.t -> 'a -> (Z.t -> 'a -> 'a) -> 'a

Fold on all integers contained in either the signed or unsigned representation of a binary.

val fold_crop_unsigned : size:Units.In_bits.t -> t -> inf:Z.t -> sup:Z.t -> 'a -> (Z.t -> 'a -> 'a) -> 'a